### 2024-02-05 #6324 @editorialbot
[PRE REVIEW]: HealpixMPI.jl: an MPI-parallel implementation of the Healpix tessellation scheme in Julia
**Submitting author:** @LeeoBianchi (<a href="http://orcid.org/0009-0002-6351-5426">Leo Alessandro Bianchi</a>)
**Repository:** https://github.com/LeeoBianchi/HealpixMPI.jl
**Branch with paper.md** (empty if default branch): JOSS-paper
**Version:** v1.0.0
