### 2024-02-05 #6324 @danielskatz
👋 @LeeoBianchi - thanks for your submission. As you can see, the next step is for me to find an editor, who will then find reviewers and get the review started.  If you have any suggestions for potential reviewers, then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
