### 2024-02-05 #6324 @dfm
@danielskatz mentioned this review to me and I wanted to list some possible reviewers with expertise in healpix and parallelization (although I'm not sure about Julia experience):

- The authors of [healpy](https://github.com/healpy/healpy), e.g. `@zonca`
- The authors of [s2fft](https://github.com/astro-informatics/s2fft), e.g. `@CosmoMatt` or `@jasonmcewen`
