### 2024-02-06 #6324 @danielskatz
> Thank you @danielskatz. I have looked at the list you provided, there are many options and I don't think I know anyone in there. For example [Yufeng Luo](https://reviewers.joss.theoj.org/reviewers/2626) is the only one that comes up when cross searching for Julia and parallel computing.

There's no reason just to suggest people you know. Anyone who seems like a potentially good reviewer based on what information available would be good to suggest, and would help the editor.
