### 2024-02-08 #6324 @danielskatz
👋 @prashjha - a ping on [the editor invitation to you](https://github.com/openjournals/joss-reviews/issues/6324#issuecomment-1927437914)...  I'm not trying to bug you, but since you are often very quickly responsive, I thought I would try again to see if this just slipped by you.
