### 2024-03-09 #6324 @prashjha
Hi @amartinhuertas, @baxmittens, @sloede, @ManuelBehrendt, @crbinz, @marcobonici, would you be interested in reviewing the JOSS submission above ([pdf link](https://github.com/openjournals/joss-papers/blob/joss.06324/joss.06324/10.21105.joss.06324.pdf))? The corresponding software is at this [link](https://github.com/LeeoBianchi/HealpixMPI.jl).

This [link](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) covers almost everything about the review process; I will be happy to help with more information.

Please let me know if you are interested. Thank you!
