### 2024-03-14 a9cac76
Update paper.md
I just ran Grammarly on this paper.
Be very careful with accepting the changes.
