### 2024-03-14 #6467 @baxmittens
Hello @LeeoBianchi

is there a way I can confirm the performance claims you have made in this paper?
Can you guide me to a script, that I can run on my servers?

Greetz max
