### 2024-03-18 #6467 @danielskatz
Hi - as the track editor, I'm just stepping in briefly to suggest that this script might be added to the repo, and pointed to from the README or elsewhere.
