### 2024-04-03 #6467 @baxmittens
I have run the script on 1-3 MPI processes

HpixMPI_adj_full_Nside4096_1MPI_32threads.txt
61.425054444
67.324124836
66.981430832
68.79545176
73.325745312

HpixMPI_a2m_full_Nside4096_1MPI_32threads.txt
105.384205176
113.553485328
114.497209256
116.934635904
122.462031616

HpixMPI_adj_full_Nside4096_2MPI_32threads.txt
63.378890184
67.107265688
81.426459472
92.957265344
120.978923648

HpixMPI_a2m_full_Nside4096_2MPI_32threads.txt
106.813835852
114.115288024
129.216200368
150.261070752
170.98185888

HpixMPI_adj_full_Nside4096_3MPI_32threads.txt
64.932011886
70.056360972
83.512326096
102.962028768
151.741963968

HpixMPI_a2m_full_Nside4096_3MPI_32threads.txt
107.305189548
112.57675122
128.54376792
146.44734288
189.411806112

HpixMPI_adj_full_Nside4096_4MPI_32threads.txt
67.3199592
70.524248256
83.624625696
104.480749696
173.429741056

HpixMPI_a2m_full_Nside4096_4MPI_32threads.txt
107.487257936
112.815567168
130.402578976
146.577254784
197.771545088
