### 2024-04-04 #6467 @LeeoBianchi
Hi @baxmittens, I can see you managed to run the benchmark scripts.

The plots you have obtained look comparable with mine on the paper, however I can see a quite steep jump between the 24(?) and 32-threads cases, especially in the adjoint transform.
This can be due to a few things. I'd be curious to know a few details about your set-up:
- How many physical computing cores do your machines have?
- What kind of connection is being used between the nodes?
- Are the 4 machines identical to each other, at least in terms of CPU and connection?

Cheers,
Leo
