### 2024-04-04 #6467 @LeeoBianchi
> I just ran them locally on one node. That is ok, since I do not want to test MPI, right?

Okay I see. Actually the way I conducted the benchmarks was to assign one MPI task to each node, so my benchmark also implied some time being spent for sending data over the cluster structure. That might explain your timings being a bit faster. Your test, however, is a good way to simulate what I have done, yes.

> My System is
>
> * 2 x AMD EPYC 7763 64-Core Processor (á 64 Cores, Hyperthreading deactivated)
> * 4020 GB RAM
>
> Therefore my timings are, as far as I can see, a little bit faster compared to yours. I dunno if both CPUs have fast access to all the memory. Maybe therefore there is a jump in the wall time?

I agree, the jump in the wall time could really look like a memory access matter to me, especially when using **all** the cores available at the same time.

Let me know if you have any further question to proceed with the review.

Cheers,
Leo
