### 2024-04-05 #6467 @baxmittens
@LeeoBianchi

Did you notice that I commited a pull request for your paper.md?
https://github.com/LeeoBianchi/HealpixMPI.jl/pull/7/commits/a9cac762b87ee55f44652214e7918b3ee09a04d5
I ran Grammarly on your paper. It has found a few commas missing and some other stuff. You can consider accepting some of these suggestions, but you don't have to. I found them quite helpful. But be careful since they were machine-generated.
