### 2024-04-08 #6467 @baxmittens
@LeeoBianchi
I did not find Community guidelines

```
Community guidelines: Are there clear guidelines for third parties wishing to

1) Contribute to the software
2) Report issues or problems with the software
3) Seek support
```
