### 2024-04-13 #6467 @LeeoBianchi
> @LeeoBianchi I added a couple of comments, opening github issues as described here. Additionally, would you mind expanding on the state-of-the-art for distributed SHT? I think I saw something based on Jax that was also GPU compatible, but I am not sure. THank you in advance! At the end of this week I should be able to run your code, so I can finish my review.

By expanding do you mean to provide some more details here, or to add something in the paper?
