### 2024-04-13 #6467 @LeeoBianchi
> @LeeoBianchi I did not find Community guidelines
>
> ```
> Community guidelines: Are there clear guidelines for third parties wishing to
>
> 1) Contribute to the software
> 2) Report issues or problems with the software
> 3) Seek support
> ```

I think you're right. I gave for granted the standard GitHub procedures/tools would have been used. Should I specify that on the Readme of the repository?
