### 2024-04-15 #6467 @baxmittens
> I think you're right. I gave for granted the standard GitHub procedures/tools would have been used. Should I specify that on the Readme of the repository?

I think that's what the JOSS guidelines prescribe. Just take any other Joss-reviewed repository like

https://github.com/gdalle/HiddenMarkovModels.jl

as a reference.
