### 2024-04-15 0e38b16
update README to include contribution and support giudelines
