### 2024-04-15 #6467 @LeeoBianchi
> > I think you're right. I gave for granted the standard GitHub procedures/tools would have been used. Should I specify that on the Readme of the repository?
>
> I think that's what the JOSS guidelines prescribe. Just take any other Joss-reviewed repository like
>
> https://github.com/gdalle/HiddenMarkovModels.jl
>
> as a reference.

Please find the updates to README.md in [0e38b16](https://github.com/LeeoBianchi/HealpixMPI.jl/commit/0e38b161b4e83c0b0a0a5cb767fb8582e7361923).

Leo
