### 2024-04-23 #6467 @baxmittens
> Hi - as the track editor, I'm just stepping in briefly to suggest that this script might be added to the repo, and pointed to from the README or elsewhere.

@LeeoBianchi Did you add the script to the repo?
Can you point me to it?
