### 2024-04-23 #6467 @LeeoBianchi
@baxmittens oops, I had changed everything locally but still haven't pushed since. Done now. You can find the [example](https://github.com/LeeoBianchi/HealpixMPI.jl/tree/JOSS-paper/examples) folder with the script inside. I also mentioned it in the readme.
