### 2024-04-25 #6467 @marcobonici
I am fine with the paper. The only missing thing from my side, is a citation that I asked @LeeoBianchi to add. After this, we are good to go!
