### 2024-05-06 66ae6cd
fix Euclid paper citations
