+++ paper/paper.bib
@@ -91,7 +91,7 @@ archivePrefix = {ascl},
 
 @misc{euclidcollaboration2023euclid,
       title={Euclid preparation: XXVIII. Modelling of the weak lensing angular power spectrum}, 
-      author={Euclid Collaboration and A. C. Deshpande and T. Kitching and A. Hall and M. L. Brown and N. Aghanim and L. Amendola and N. Auricchio and M. Baldi and R. Bender and D. Bonino and E. Branchini and M. Brescia and J. Brinchmann and S. Camera and G. P. Candini and V. Capobianco and C. Carbone and V. F. Cardone and J. Carretero and F. J. Castander and M. Castellano and S. Cavuoti and A. Cimatti and R. Cledassou and G. Congedo and C. J. Conselice and L. Conversi and L. Corcione and F. Courbin and M. Cropper and A. Da Silva and H. Degaudenzi and M. Douspis and F. Dubath and C. A. J. Duncan and X. Dupac and S. Farrens and S. Ferriol and P. Fosalba and M. Frailis and E. Franceschi and M. Fumana and S. Galeotta and B. Garilli and B. Gillis and C. Giocoli and A. Grazian and F. Grupp and S. V. H. Haugan and H. Hoekstra and W. Holmes and A. Hornstrup and P. Hudelot and K. Jahnke and S. Kermiche and M. Kilbinger and M. Kunz and H. Kurki-Suonio and S. Ligori and P. B. Lilje and I. Lloro and E. Maiorano and O. Mansutti and O. Marggraf and K. Markovic and F. Marulli and R. Massey and S. Mei and Y. Mellier and M. Meneghetti and G. Meylan and L. Moscardini and S. -M. Niemi and J. W. Nightingale and T. Nutma and C. Padilla and S. Paltani and F. Pasian and K. Pedersen and V. Pettorino and S. Pires and G. Polenta and M. Poncet and L. A. Popa and F. Raison and A. Renzi and J. Rhodes and G. Riccio and E. Romelli and M. Roncarelli and E. Rossetti and R. Saglia and D. Sapone and B. Sartoris and P. Schneider and T. Schrabback and A. Secroun and G. Seidel and S. Serrano and C. Sirignano and G. Sirri and L. Stanco and P. Tallada-Crespi and I. Tereno and R. Toledo-Moreo and F. Torradeflot and I. Tutusaus and E. A. Valentijn and L. Valenziano and T. Vassallo and Y. Wang and J. Weller and A. Zacchei and G. Zamorani and J. Zoubian and S. Andreon and S. Bardelli and A. Boucaud and E. Bozzo and C. Colodro-Conde and D. Di Ferdinando and G. Fabbian and M. Farina and J. Gracia-Carpio and E. Keihanen and V. Lindholm and N. Mauri and V. Scottez and M. Tenti and E. Zucca and Y. Akrami and C. Baccigalupi and A. Balaguera-Antolinez and M. Ballardini and F. Bernardeau and A. Biviano and A. Blanchard and A. S. Borlaff and C. Burigana and R. Cabanac and A. Cappi and C. S. Carvalho and S. Casas and G. Castignani and T. Castro and K. C. Chambers and A. R. Cooray and J. Coupon and H. M. Courtois and S. Davini and S. de la Torre and G. De Lucia and G. Desprez and H. Dole and J. A. Escartin and S. Escoffier and I. Ferrero and F. Finelli and J. Garcia-Bellido and K. George and F. Giacomini and G. Gozaliasl and H. Hildebrandt and J. J. E. Kajava and V. Kansal and C. C. Kirkpatrick and L. Legrand and A. Loureiro and J. Macias-Perez and M. Magliocchetti and G. 
[…]
