### 2024-05-09 #6467 @prashjha
@LeeoBianchi, while checking references, I noticed some citations appear to be missing DOIs. Can you check?
