+++ paper/paper.md
@@ -91,7 +91,7 @@ In addition, refer to [Jommander](https://github.com/LeeoBianchi/Jommander.jl), 
 This section shows the results of parallel benchmark tests conducted on `HealpixMPI.jl`.
 In particular, a strong-scaling scenario is analyzed: given a problem of fixed size, the wall time improvement is measured as the number of cores exploited in the computation is increased.
 
-To obtain a reliable measurement of massively parallel spherical harmonics wall time is certainly nontrivial, especially for tests implying a high number of cores; intermittent operating system activity can significantly distort the measurement of short time scales.
+To obtain a reliable measurement of massively parallel spherical harmonics wall time is certainly nontrivial, especially for tests employing a high number of cores; intermittent operating system activity can significantly distort the measurement of short time scales.
 For this reason, the benchmark tests were carried out by timing a batch of 20 `alm2map` + `adjoint_alm2map` SHT pairs.
 For reference, the scaling shown here is relative to unpolarized spherical harmonics with $\mathrm{N}_\mathrm{side} = 4096$ and $\ell_{\mathrm{max}} = 12287$ and were carried out on the [Hyades cluster](https://www.mn.uio.no/astro/english/services/it/help/basic-services/compute-resources.html) of the University of Oslo.
 The benchmark results are quantified as the wall time multiplied by the total number of cores, shown in a 3D plot (\autoref{fig:bench}) as a function of the number of local threads and MPI tasks (always one per node).
