### 2024-05-13 #6467 @LeeoBianchi
Hi @prashjha, I have updated the paper according to what you suggested. For what concerns Zenodo, should I just upload the entire repo as a .zip? It also asks for a DOI which I of course don't have yet, should I just go on without it?
