### 2024-05-13 #6467 @baxmittens
@LeeoBianchi Just let zenodo generate a doi for you. the zenodo doi and joss doi are different things. (I am like 80% sure...)
