+++ paper/paper.bib
@@ -14,7 +14,7 @@ archivePrefix = {ascl},
 }
 
 @article{Watts_2023,
-   title={COSMOGLOBE DR1 results: I. Improved Wilkinson Microwave Anisotropy Probe maps through Bayesian end-to-end analysis},
+   title={COSMOGLOBE DR1 results: I. Improved {W}ilkinson Microwave Anisotropy Probe maps through {B}ayesian end-to-end analysis},
    volume={679},
    ISSN={1432-0746},
    url={http://dx.doi.org/10.1051/0004-6361/202346414},
@@ -49,7 +49,7 @@ archivePrefix = {ascl},
 	number = {2},
 	pages = {227--241},
 	author = {H. K. Eriksen and I. J. O'Dwyer and J. B. Jewell and B. D. Wandelt and D. L. Larson and K. M. Gorski and S. Levin and A. J. Banday and P. B. Lilje},
-	title = {Power Spectrum Estimation from High-Resolution Maps by Gibbs Sampling},
+	title = {Power Spectrum Estimation from High-Resolution Maps by {G}ibbs Sampling},
 
 	journal = {The Astrophysical Journal Supplement Series}
 }
