### 2024-05-17 #6467 @danielskatz
@LeeoBianchi - As track editor, I've proofread the paper and suggest the changes in https://github.com/LeeoBianchi/HealpixMPI.jl/pull/10.

In addition, something is wrong with the grammar of the second sentence of the summary, and I can't suggest a fix because I'm a little unsure what the intent of the sentence is.  Another issue is that SHT sometimes is used for SHTs, and sometimes SHTs is used. I suggest using SHT for one, and SHTs for multiple, consistently through the paper.

Please let me know when the PR is merged (or what you disagree with), and when you have addressed the other two points, then we can proceed to publication.
