+++ paper/paper.md
@@ -24,15 +24,15 @@ bibliography: paper.bib
 
 # Summary
 
-Spherical Harmonic Transforms (SHT) can be seen as Fourier Transforms' spherical, two-dimensional counterparts, casting real-space data to the spectral domain and vice versa.
-As in Fourier analysis a function is decomposed into a set of amplitude coefficients, through an SHT, any spherically-symmetric field defined in real space can be decomposed into a set of complex harmonic coefficients $a_{\ell, m}$, commonly referred to as alms, each quantifying the contribution of the corresponding spherical harmonic function.
+Spherical Harmonic Transforms (SHTs) can be seen as Fourier Transforms' spherical, two-dimensional counterparts, casting real-space data to the spectral domain and vice versa.
+As in Fourier analysis a function is decomposed into a set of amplitude coefficients, an SHT allows to decompose any spherically-symmetric field, defined in real space, into a set of complex harmonic coefficients $a_{\ell, m}$, commonly referred to as alms, each quantifying the contribution of the corresponding spherical harmonic function.
 
 SHTs are important for a wide variety of theoretical and practical scientific applications, including particle physics, astrophysics, and cosmology.
-However, the SHTs are generally computationally expensive operations and thus often constitute the *bottleneck* of the scientific software they are part of.
-For this reason, much effort has been spent over the last couple of decades to obtain fast and efficient SHT implementations.
+However, SHTs are generally computationally expensive operations and thus often constitute the *bottleneck* of the scientific software they are part of.
+For this reason, much effort has been spent over the last couple of decades to obtain fast and efficient SHTs implementations.
 In such a setting, parallel computing naturally comes into play, especially for time-consuming software to be run on large High-Performance Computing (HPC) clusters.
 
-The Julia package `HealpixMPI.jl` constitutes an extension package of `Healpix.jl` [@Healpix_jl], efficiently parallelizing its SHT functionalities.
+The Julia package `HealpixMPI.jl` constitutes an extension package of `Healpix.jl` [@Healpix_jl], efficiently parallelizing its SHT-related functionalities.
 `Healpix.jl` is a Julia-only implementation of the HEALPix [@HEALPix] library, which provides one of the most used two-sphere tessellation schemes and a series of SHTs-related functions.
 
 The main goal of the Julia package presented in this paper, `HealpixMPI.jl`, is to efficiently employ a large number of computing cores to perform fast spherical harmonic transforms.
@@ -46,7 +46,7 @@ This paper presents the key features implemented to achieve this, together with 
 # Statement of need
 
 Together with a variety of applications, spherical harmonic transforms are extremely relevant in different cosmological research topics, e.g., @Loureiro_2023 and @euclidcollaboration2023euclid.
-Among those, SHT are essential for the analysis of cosmic microwave background (CMB) radiation, which is one of the most active cosmology research areas.
+Among those, SHTs are essential for the analysis of cosmic microwave background (CMB) radiation, which is one of the most active cosmology research areas.
 CMB radiation is, in fact, very conveniently described as a temperature (and polarization) field on the celestial sphere, making spherical harmonics the most natural mathematical tool for analyzing its measured signal.
 On the other hand, from a computational point of view, CMB field measurements need, of course, to be discretized, requiring a mathematically consistent pixelization of the sphere and the functions defined on it.
 This is exactly the goal HEALPix was targeting when it was released more than two decades ago; it quickly became the standard library for CMB numerical analysis.
@@ -70,7 +70,7 @@ In fact, `DUCC`’s code is derived directly from `libsharp`, but has been signi
 
[…]
