### 2024-05-17 #6467 @LeeoBianchi
> @LeeoBianchi - As track editor, I've proofread the paper and suggested changes in [LeeoBianchi/HealpixMPI.jl#10](https://github.com/LeeoBianchi/HealpixMPI.jl/pull/10).
>
> In addition, something is wrong with the grammar of the second sentence of the summary, and I can't suggest a fix because I'm a little unsure what the intent of the sentence is. Another issue is that SHT sometimes is used for SHTs, and sometimes SHTs is used. I suggest using SHT for one, and SHTs for multiple, consistently through the paper.
>
> Please let me know when the PR is merged (or what you disagree with), and when you have addressed the other two points, then we can proceed to publication.

Hi, thank you for proofreading my paper. I merged your PR as all the changes proposed looked good.

I tried to rephrase the sentence you pointed out, I hope is more clear now.

You were right about the SHT/SHTs issue, I think grammarly messed up most of them, I should have fixed that too.

I pushed alle the fixes in a new [commit](https://github.com/LeeoBianchi/HealpixMPI.jl/commit/d0f3726108aefb17863dbeb3670d7115ba9f468b), let me know I should fix anything else.

Cheers
