### 2024-05-18 #6467 @LeeoBianchi
> I've added some comments in your commit

I should have implemented all the proposed changes. Thank you!
