### 2024-05-19 #6467 @danielskatz
I've added one more small comment on https://github.com/LeeoBianchi/HealpixMPI.jl/commit/dc177d0de2422d3abb2dd9557e61f08a31d8b538 - once this is fixed, I think we'll be ready to go.
