### 2024-05-20 e4d3384
Fix typo in paper.md
